third_party/blink/web_tests/external/wpt/webrtc/RTCPeerConnection-setLocalDescription-offer-expected.txt

This is a testharness.js-based test.
PASS setLocalDescription with valid offer should succeed
PASS setLocalDescription with type offer and null sdp should use lastOffer generated from createOffer
PASS setLocalDescription() with offer not created by own createOffer() should reject with InvalidModificationError
FAIL Set created offer other than last offer should reject with InvalidModificationError assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Creating and setting offer multiple times should succeed
PASS Setting previously generated offer after a call to createAnswer should work
PASS Negotiation works when there has been a repeated setLocalDescription(offer)
FAIL setLocalDescription(offer) should update internal state with a queued task, in the right order assert_equals: pendingRemoteDescription should never be set due to sLD expected null but got object "[object RTCSessionDescription]"
Harness: the test ran to completion.
